requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pdf2image==1.16.0
Pillow==8.3.1
prettytable==2.1.0
python-irodsclient==1.0.0
six==1.16.0
wcwidth==0.2.5
Werkzeug==2.0.1
xmlrunner==1.7.7
